auth-service/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1
python-multipart==0.0.6

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client
httpx==0.25.2

# OpenTelemetry dependencies
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-zipkin-json==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-semantic-conventions==0.42b0

# In a real application
# SQLAlchemy==2.0.23
# psycopg2-binary==2.9.9
# pika==1.3.2
# For RabbitMQ

# HTTP client for health checks
requests==2.31.0

pytest==7.4.3
pytest-asyncio==0.21.1
